feat: move workflow/metadata models to `baseinvocation.py`
needed to prevent circular imports
