+++ invokeai/app/services/image_files/image_files_disk.py
@@ -7,9 +7,8 @@ from PIL import Image, PngImagePlugin
 from PIL.Image import Image as PILImageType
 from send2trash import send2trash
 
-from invokeai.app.invocations.metadata import MetadataField
+from invokeai.app.invocations.baseinvocation import MetadataField, WorkflowField
 from invokeai.app.services.invoker import Invoker
-from invokeai.app.services.workflow_records.workflow_records_common import WorkflowField
 from invokeai.app.util.thumbnails import get_thumbnail_name, make_thumbnail
 
 from .image_files_base import ImageFileStorageBase
