+++ invokeai/app/services/images/images_default.py
@@ -2,10 +2,9 @@ from typing import Optional
 
 from PIL.Image import Image as PILImageType
 
-from invokeai.app.invocations.metadata import MetadataField
+from invokeai.app.invocations.baseinvocation import MetadataField, WorkflowField
 from invokeai.app.services.invoker import Invoker
 from invokeai.app.services.shared.pagination import OffsetPaginatedResults
-from invokeai.app.services.workflow_records.workflow_records_common import WorkflowField
 
 from ..image_files.image_files_common import (
     ImageFileDeleteException,
